closer, but suffering from media-type conflicts
git-svn-id: http://subversion.assembla.com/svn/Debrief@3190 cb33b658-6c9e-41a7-9690-cba343611204

+++ trunk/org.mwc.asset.comms/src/org/mwc/asset/comms/restlet/data/ScenarioStateResource.java
@@ -10,5 +10,5 @@ public interface ScenarioStateResource {
 //	@Post
 //	public void accept(String event, long time, String description);
 	@Post
-	public void accept(int event, String val2);
+	public void accept(long event, String val2);
 }
